% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvgk.R
\name{mvgk}
\alias{mvgk}
\title{mvgk}
\usage{
mvgk(ED)
}
\arguments{
\item{ED}{A distance matrix in the form of a symmetric numeric matrix. Row-
and column-order must match the order of the levels of the environment
factor in the dataframe that will be passed to ASReml-R.}
}
\value{
A function compatible with asreml::own() that uses the distance matrix
to compute a non-linear kernel given a bandwidth \code{h}. This non-linear
kernel is then combined with unique variances for each environment to compute
the full covariance matrix and its partial derivatives.
}
\description{
mvgk
}
