% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlk.R
\name{mvlk}
\alias{mvlk}
\title{mvlk}
\usage{
mvlk(C)
}
\arguments{
\item{C}{A linear kernel in the form of a symmetric numeric matrix. Row-
and column-order must match the order of the levels of the environment
factor in the dataframe that will be passed to ASReml-R.}
}
\value{
A function compatible with asreml::own() that uses the linear kernel
and a unique variance for each environment to compute the full covariance
matrix and its partial derivatives.
}
\description{
The \code{mvlk} function computes the covariance matrix and partial derivative matrices
for the multiple-variance linear kernel model (MV-LK).
This function can be used for both GxE and GxExM/GxExT models.
The MV-LK model assumes that every environment by management (or environment by trait) combination has a
unique genetic variance. For GxE data, this means that the number of genetic
variances that are estimated is equal to the number of environments.
}
\examples{
\dontrun{
# For GxE data:
vf <- cornfruit::mvlk(C = linearKernel)
q <- length(levels(df$Environment))
# Variances:
init <- rep(0.1, q)
type <- rep("V", q)
cons <- rep("P", q)
fit <- asreml(fixed = Y ~ 1 + Environment,
              random = ~ own(Environment, "vf", init, type, cons):vm(Genotype, K),
              residual = ~ units,
              data = df)

# For GxExM data:
vf <- cornfruit::mvlk(C = linearKernel)
p <- length(levels(df$Management))
q <- length(levels(df$Environment))
# Variances + correlations:
init <- c(rep(0.1, p * q), rep(0.1, (p^2 - p) / 2))
type <- c(rep("V", p * q), rep("R", (p^2 - p) / 2))
cons <- c(rep("P", p * q), rep("U", (p^2 - p) / 2))
fit <- asreml(fixed = Y ~ 1 + ManagementEnvironment,
              random = ~ own(ManagementEnvironment, "vf", init, type, cons):vm(Genotype, K),
              residual = ~ units,
              data = df)

# For GxExT data:
vf <- cornfruit::mvlk(C = linearKernel)
p <- length(levels(df$Trait))
q <- length(levels(df$Environment))
# Variances + correlations:
init <- c(rep(0.1, p * q), rep(0.1, (p^2 - p) / 2))
type <- c(rep("V", p * q), rep("R", (p^2 - p) / 2))
cons <- c(rep("P", p * q), rep("U", (p^2 - p) / 2))
fit <- asreml(fixed = Y ~ 1 + TraitEnvironment,
              random = ~ own(TraitEnvironment, "vf", init, type, cons):vm(Genotype, K),
              residual = ~ units,
              data = df,
              asmv = Trait)
}

}
