% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlk.R
\name{mvlk}
\alias{mvlk}
\title{mvlk}
\usage{
mvlk(EC)
}
\arguments{
\item{EC}{A linear kernel in the form of a symmetric numeric matrix. Row-
and column-order must match the order of the levels of the environment
factor in the dataframe that will be passed to ASReml-R.}
}
\value{
A function compatible with asreml::own() that uses the linear kernel
and a unique variance for each environment to compute the full covariance
matrix and its partial derivatives.
}
\description{
mvlk
}
